% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregating.R
\name{aggregate_rp5}
\alias{aggregate_rp5}
\title{Агрегация данных о температуре, количестве осадков и высоте снежного покрова}
\usage{
aggregate_rp5(df, resolution)
}
\arguments{
\item{df}{a tibble: тиббл, полученный после выполнения функций \code{\link{read_rp5_csv}} или \code{\link{read_rp5_folder}} и
\code{\link{separate_date_rp5}}. Должен содержать в себе столбцы: обязательно - Year, Month, Day;
опционально - любой из T, RRR, sss или в любом сочетании.
Для других столбцов агрегация производиться не будет, даже если они присутствуют в датафрейме.}

\item{resolution}{character: до какого разрешения агрегировать данные? До суточного: resolution = 'day',
месячного: resolution = 'month', годового: resolution = 'year'.}
}
\value{
a tibble: Тиббл со столбцами Year, Month, Day (в зависимости от выбранного разрешения),
N - количество наблюдений в группе (обязательно), и, в зависимости от исходного набора столбцов
T_avg - средняя температура, RRR_sum - сумма осадков, sss_avg - средняя высота снежного покрова.
}
\description{
Вычисляет среднюю температуру, сумму осадков и среднюю высоту снежного покрова для заданного разрешения: годового,
месячного или суточного. Пропущенные значения исключает из расчетов. Работает с тибблом - результатом
последовательного выполнения функций \code{\link{read_rp5_csv}} или \code{\link{read_rp5_folder}} и
\code{\link{separate_date_rp5}}.
}
\examples{
\dontrun{
aggregate_rp5(df, 'day')}
}
