% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_transitions.R
\name{find_transitions}
\alias{find_transitions}
\title{Найти переход заданного параметра через пороговое значение}
\usage{
find_transitions(df, parameter, threshold)
}
\arguments{
\item{df}{тиббл, содержащий как минимум 1 столбец - параметр, для которого будут вычисляться переходы (например, с температурой). Тиббл предварительно можно сгруппировать (например, вычислить переходы параметра через порог для каждого года, месяца и дня)}

\item{parameter}{a character: название столбца, для которого будут вычисляться переходы через пороговое значение threshold}

\item{threshold}{a numeric: пороговое значение}
}
\value{
a tibble: содержит группирующие столбцы, столбец, для которого вычислялись переходы, а также столбец chng - порядковый номер перехода параметра через заданное пороговое значение.
}
\description{
Служебная функция.
Вычисляет факт перехода необходимого параметра через заданное пороговое значение.
Результат используется другими функциями внутри пакета.
По умолчанию удаляет пропущенные значения в параметре, поэтому NA предварительно необходимо обрабатывать.
}
\examples{
\dontrun{
Найти переходы температур через 0 градусов
find_transitions(dataframe, T_avg, 0)}
}
