% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_transitions.R
\name{freq_of_transitions}
\alias{freq_of_transitions}
\title{Частота переходов среднесуточного параметра через пороговое значение (в месяц)}
\usage{
freq_of_transitions(df, parameter, threshold)
}
\arguments{
\item{df}{a tibble: датафрейм со столбцами Year, Month, Параметр (например, T_avg). (опционально Day. Необходимо использовать датафрейм суточного разрешения, то есть одно измерение параметра должно соответствовать одним суткам)}

\item{parameter}{a character: название столбца - интересующего параметра}

\item{threshold}{numeric: пороговое значение, через которое совершаются переходы}
}
\value{
a tibble: датафрейм со столбцами:
Year
Month
Num_of_transitions - количество переходов через заданный порог
days_in_month - количество дней в месяце (рассчитывается по входным данным)
Freq_of_transitions = Num_of_transitions/days_in_month  : частота переходов (переходов в сутки)
}
\description{
Рассчитывает частоту как количество переходов параметра через заданное пороговое значение, поделенное на продолжительность месяца (в днях)
}
\examples{
\dontrun{
# пример полного папйплайна
read_rp5_folder('path') |>    # считать все файлы в папке
    separate_date_rp5() |>    # разбить даты на отдельные столбцы
    na_fill_temp_rp5() |>     # заполнить пропуски в исходных значениях температуры
    aggregate_rp5('day') |>   # агрегировать данные до суточного разрешения
    na_fill_sn_rp5(10) |>     # заполнить пропуски в измерениях глубин снежного покрова
    freq_of_transitions(T_avg, 0)} # рассчитать частоту и количество переходов среднесуточных температур через 0
}
