% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_with_trends}
\alias{graph_with_trends}
\title{График трендов}
\usage{
graph_with_trends(
  data,
  trend,
  devide_by_month = FALSE,
  y,
  y_smoothed,
  geom_type = "line",
  x_label = "Years",
  y_label = "Temperature, °C"
)
}
\arguments{
\item{data}{a tibble: тиббл с агрегированными данными. Должен содержать столбцы Year, (Month - для данных месячного разрешения), а также столбец с визуализируемым параметром}

\item{trend}{a tibble: тиббл с параметрами тренда.}

\item{devide_by_month}{logical, default = FALSE: если данные месячного разрешения, установить TRUE, чтобы график разбивался на фасетки по месяцам.}

\item{y}{название столбца с параметром для визуализации}

\item{y_smoothed}{название столбца со сглаженным параметром для визуализации. Если вы не использовали сглаживание, укажите то же имя столбца, что и в y.}

\item{geom_type}{тип геометрии. По умолчанию 'line', возможен вариант 'col'}

\item{x_label}{подпись оси x. По умолчанию 'Years'}

\item{y_label}{подпись оси y. По умолчанию 'Temperature, °C'}
}
\value{
ggplot2 график
}
\description{
График трендов для данных месячного или годового разрешения
}
\examples{
\dontrun{
graph_with_trends(data = monthly,
                 trend = trends_by_month,
                 devide_by_month = TRUE,
                 y = T_avg,
                 y_smoothed = Tavg_smoothed,
                 geom_type = 'line',
                 x_label = 'Годы',
                 y_label = "температуры")
graph_with_trends(data = annualy,
                 trend = annualy_trend,
                 y = T_avg,
                 y_smoothed = T_avg,
                 geom_type = 'col')}
}
