% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{linear_trend}
\alias{linear_trend}
\title{Расчет параметров линейного тренда}
\usage{
linear_trend(df, x, y)
}
\arguments{
\item{df}{a tibble: источник данных.}

\item{x}{предиктор (в текущем контексте это временная метка, например, год. Если год в данных один - используйте месяц, если он один - дату и т.д.)}

\item{y}{заисимая переменная (в текущем контексте это чаще всего сглаженный параметр, например средняя температура, сглаженная скользящим средним)}
}
\value{
a tibble: датафрейм со столбцами:
r2 - коэффициент детерминации
p-value - для угла наклона регрессионной прямой (коэффииент b, или slope)
intercept - независимый член регрессионного уравнения
equation - уравнение линейного тренда
Опционально добавится столбец с группирующей переменной
}
\description{
Расчет линейного тренда для сглаженного параметра и создание уравнения тренда.
Используется для нанесения уравнения на график, а также для построения сводной таблицы результатов анализа трендов.
Можно применять как к группированному датафрейму (линейный тренд будет вычислен для каждой группы отдельно),
так и без группировки.
}
\examples{
\dontrun{
# простое применение
linear_trend(your_data, Year, Tavg_smoothed)

# Шаблон полного пайплайна:
read_rp5_folder('inst/extdata/') |>   # путь к файлам
  separate_date_rp5() |>              # разбить дату на отдельные столбцы
  na_fill_temp_rp5() |>               # заполнить пропуски в температурах
  aggregate_rp5('day') |>             # агрегировать до суточного разрешения
  na_fill_sn_rp5(10) |>               # заполнить пропуски в измерениях снега
  aggregate_rp5('month') |>           # агрегировать до месячного разрешения
  dplyr::mutate(Tavg_smoothed = means_smooth(T_avg)) |>   # применить сглаживание скользящим средним
  linear_trend(Year, Tavg_smoothed)   # рассчитать линейный тренд для сглаженного параметра

# расчет тренда, например, январских температур за все имеющиеся годы
monthly |> # данные месячного разрешения (после шага aggregate_rp5('month') из предыдущего примера)
  dplyr::filter(Month == 1) |>
  linear_trend(Year, Tavg_smoothed)

# расчет тренда внутри одного года:
monthly |>
  dplyr::filter(Year == 2023) |>
  linear_trend(Month, Tavg_smoothed)


# расчет тренда для каждого года в исходном датафрейме
 monthly |>
   dplyr::group_by(Year) |>
   linear_trend(Month, Tavg_smoothed)
   }
}
