% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{means_smooth}
\alias{means_smooth}
\title{Расчет скользящего среднего}
\usage{
means_smooth(column, window = 10)
}
\arguments{
\item{column}{: столбец для сглаживания}

\item{window}{: ширина окна}
}
\value{
vector: вектор, который можно сохранить в отдельный столбец
}
\description{
Расчет скользящего среднего
}
\examples{
\dontrun{
df |>
  means_smooth(T_avg)
  }
}
