% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{monthly_graph}
\alias{monthly_graph}
\title{График трендов параметра для каждого месяца}
\usage{
monthly_graph(
  monthly_data,
  monthly_trend,
  y,
  y_smoothed,
  geom_type = "line",
  x_label,
  y_label
)
}
\arguments{
\item{monthly_data}{a tibble: тиббл с данными месячного разрешения. Должен содержать столбцы Year, Month, а также столбец с интересующим параметром и его сглаженной версией}

\item{monthly_trend}{a tibble: тиббл с параметрами линейного тренда для каждого месяца (результат функции \code{\link{linear_trend}})}

\item{y}{: название столбца с параметром (например, T_avg)}

\item{y_smoothed}{: сглаженный параметр (например, Tavg_smoothed)}

\item{geom_type}{: тип геометрии. По умолчанию 'line' - линейный график. Другой вариант - col - столбчатый график.}

\item{x_label}{: подпись оси Х.}

\item{y_label}{: подпись оси Y.}
}
\value{
ggplot2 график
}
\description{
График трендов параметра для каждого месяца
}
\examples{
\dontrun{
monthly_graph(monthly, trends_by_month, T_avg, Tavg_smoothed)
}

}
