% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_processing.R
\name{na_fill_sn_rp5}
\alias{na_fill_sn_rp5}
\title{Обработка пропущенных значений глубины снежного покрова (только данные суточного разрешения)}
\usage{
na_fill_sn_rp5(df, threshold, replace_original = TRUE)
}
\arguments{
\item{df}{a tibble: данные суточного разрешения со столбцами Year, Month, Day и sss_svg.}

\item{threshold}{numeric: пороговое значение длины "окна" с пропусками}

\item{replace_original}{logical: заменять ли исходный столбец? по умолчанию TRUE}
}
\value{
a tibble: исходный тиббл, где пропуски в sss_avg заменены (либо добавлен новый столбец sss_replaced - удобно для диагностики)
}
\description{
Логика работы: если значения глубины снежного покрова пропущены длительное время (заданный порог) - он растаял.
Такие пропуски стоит заменить на 0.
Если число подряд следующих пропусков менее порога - заменить на последнее значение перед пропусками.
Работает только с датафреймом суточного разрешения! (должны быть столбцы Year, Month, Day)
}
\examples{
\dontrun{
read_rp5_folder('path') |>
      separate_date_rp5() |>
      na_fill_temp_rp5(replace_original = F) |>
      aggregate_rp5('day') |>
      na_fill_sn_rp5(10, replace_original = F)}
}
