% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_processing.R
\name{na_fill_temp_rp5}
\alias{na_fill_temp_rp5}
\title{Заполнение пропущенных значений температур}
\usage{
na_fill_temp_rp5(df, replace_original = TRUE)
}
\arguments{
\item{df}{a tibble: Должен содержать столбцы с температурой (ожидаются названия T или T_avg)}

\item{replace_original}{LOGICAL: заменять ли исходный столбец новым столбцом без пропусков? По умолчанию TRUE}
}
\value{
a tibble:  исходный тиббл с заполненными пропусками в столбце с температурами, либо с новым столбцом, в котором заполнены пропуски (исходный столбец с температурами остается неизменным)
}
\description{
Выполняется для "сырых" данных, но применим и к агрегированным. Заполнение пропусков методом линейной интерполяции
(пакет zoo). По умолчанию пропуски в начале или конце датафрейма заполняются соседними значениями.
}
\examples{
\dontrun{
# Подсчет пропусков в агрегированных данных
df <- read_rp5_folder('path') |>
separate_date_rp5() |>
  na_fill_temp_rp5(replace_original = F) }
}
