% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_transitions.R
\name{num_of_transitions}
\alias{num_of_transitions}
\title{Посчитать количество переходов параметра через заданный порог}
\usage{
num_of_transitions(df, parameter, threshold)
}
\arguments{
\item{df}{\itemize{
\item тиббл, содержащий как минимум 1 столбец - параметр, для которого будут вычисляться переходы (например, с температурой). Тиббл предварительно можно сгруппировать (например, вычислить переходы параметра через порог для каждого года, месяца и дня)
}}

\item{parameter}{a character: название столбца, для которого будут вычисляться переходы через пороговое значение threshold}

\item{threshold}{a numeric: пороговое значение}
}
\value{
a tibble: содержит группирующие столбцы (если есть), столбец, для которого вычислялись переходы, а также столбец Num_of_transitions - количество переходов параметра через заданное значение
}
\description{
Подсчет количества переходов параметра через заданный порог.
По умолчанию удаляет пропущенные значения в параметре, поэтому NA предварительно необходимо обрабатывать.
Может применяться после группировки датафрейма, например, для расчета количества переходов в каждом месяце, году и т.д.
}
\examples{
\dontrun{
Найти переходы температур через 0 градусов
num_of_transitions(dataframe, T_avg, 0)

#' Найти переходы температур через 0 градусов для каждого месяца
dataframe |>
group_by(Year, Month) |>
num_of_transitions(T_avg, 0)}
}
