% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_rp5.R
\name{process_date_rp5}
\alias{process_date_rp5}
\title{Date column processing: output of year, month, day in separate columns}
\usage{
process_date_rp5(
  df,
  col_with_local_time = 1,
  format_of_local_time = "\%d.\%m.\%Y \%R"
)
}
\arguments{
\item{df}{a tibble: the result of executing the read_rp5_csv() or read_rp5_excel() functions}

\item{col_with_local_time}{integer: the ordinal number of the column containing information about the local time. By default : 1.}

\item{format_of_local_time}{a character: a string with the time format. By default: "\%d.\%m.\%Y \%R". Read more: https://lubridate.tidyverse.org/reference/as_date.html}
}
\value{
a tibble
}
\description{
Date column processing: output of year, month, day in separate columns
}
\examples{
\dontrun{
process_date_rp5(df, col_with_local_time = 1, format_of_local_time = "\%d.\%m.\%Y \%R")}
}
