% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rp5.R
\name{read_rp5_csv}
\alias{read_rp5_csv}
\title{Импорт csv файла, скачанного с сайта  \href{https://rp5.ru/}{rp5.ru}}
\usage{
read_rp5_csv(
  path,
  skip = 6,
  delim = ";",
  col_select = c(Local_time = 1, dplyr::matches("^T$"), dplyr::matches("^RRR$"),
    dplyr::matches("^sss$")),
  encoding = NULL,
  suppress_warnings = T
)
}
\arguments{
\item{path}{character: путь к csv файлу.}

\item{skip}{integer: количество строк в csv-файле, которые необходимо пропустить при импорте.
(обычно в верхних строках при скачивании с сайта rp5.ru хранятся метаданные).
По умолчанию \code{skip=6}}

\item{delim}{character: разделитель. По умолчанию \code{delim=';'}}

\item{col_select}{character: вектор с названиями столбцов для отбора из исходного файла.
Работает по принципу \code{\link[dplyr]{select}}. По умолчанию отбирает столбцы с местным временем,
температурой (столбец T), количеством выпавших осадков (столбец RRR) и высотой снежного покрова
столбец (sss). По стандарту rp5 местное время располагается в первом столбце,
но можно указать и номер другого столбца. Автоматически переименуется в Local_time для унификации.}

\item{encoding}{character: кодировка csv-файла.По умолчанию \code{encoding = NULL},
и функция пытается распознать кодировку, указанную в имени файла.
При скачивании csv с rp5.ru информация о кодировке записана в названии, поэтому для автоматического
распознавания кодировки лучше не переименовывать файл. Если возникают ошибки, вы можете установить
значение вручную. Варианты кодировок,
в которых можно скачать csv с сайта rp5.ru: "utf-8", unicode" или "ANSI" = "windows-1251"}

\item{suppress_warnings}{bool: по умолчанию предупреждения подавляются.
Для отладки вы можете их включить: \code{suppress_warnings = T}}
}
\value{
A tibble: тиббл со столбцами Local_time, T, RRR, sss
}
\description{
Высокоуровневая обертка для функции
\code{\link[readr]{read_delim}}, адаптированной для импорта csv файла,
скачанного с сайта \href{https://rp5.ru/}{rp5.ru}. При импорте
происходит автоматическое удаление метаданных и отбор столбцов с местным временем, температурой (T),
количеством осадков (RRR), высотой снежного покрова (sss). Cтолбцы T, RRR, sss
автоматически очищаются от лишних символов, остаются только дробные или целые числа.
}
\examples{
\dontrun{
# В самом распространенном случае вам нужно просто выполнить:
read_csv_rp5("путь к вашему файлу")


# Каждый параметр функции можно настроить:
read_csv_rp5("path/to/your/file",
              skip = 6,
              delim = ";",
              col_select =  c(Local_time = 1,
                              dplyr::matches('^T$'),
                              dplyr::matches('^sss$')),
              encoding = "UTF-8",
              suppress_warnings = F)}
}
