% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rp5.R
\name{read_rp5_folder}
\alias{read_rp5_folder}
\title{Импорт всех csv файлов из папки и объединение их в один тиббл}
\usage{
read_rp5_folder(
  path,
  skip = 6,
  delim = ";",
  col_select = c(Local_time = 1, dplyr::matches("^T$"), dplyr::matches("^RRR$"),
    dplyr::matches("^sss$")),
  encoding = NULL,
  suppress_warnings = T
)
}
\arguments{
\item{path}{character: путь к csv файлу.}

\item{skip}{integer: количество строк в csv-файле, которые необходимо пропустить при импорте.
(обычно в верхних строках при скачивании с сайта rp5.ru хранятся метаданные).
По умолчанию \code{skip=6}}

\item{delim}{character: разделитель. По умолчанию \code{delim=';'}}

\item{col_select}{character: вектор с названиями столбцов для отбора из исходного файла.
Работает по принципу \code{\link[dplyr]{select}}. По умолчанию отбирает столбцы с местным временем,
температурой (столбец T), количеством выпавших осадков (столбец RRR) и высотой снежного покрова
столбец (sss). По стандарту rp5 местное время располагается в первом столбце,
но можно указать и номер другого столбца. Автоматически переименуется в Local_time для унификации.}

\item{encoding}{character: кодировка csv-файла.По умолчанию \code{encoding = NULL},
и функция пытается распознать кодировку, указанную в имени файла.
При скачивании csv с rp5.ru информация о кодировке записана в названии, поэтому для автоматического
распознавания кодировки лучше не переименовывать файл. Если возникают ошибки, вы можете установить
значение вручную. Другие варианты кодировок,
в которых можно скачать csv с сайта rp5.ru: "utf-8", unicode" или "ANSI" = "windows-1251"}

\item{suppress_warnings}{bool: по умолчанию предупреждения подавляются.
Для отладки вы можете их включить: \code{suppress_warnings = T}}
}
\value{
tibble со столбцами, указанными в col_select
}
\description{
Обертка над функцией \code{\link{read_rp5_csv}}, позволяющая импортировать сразу все csv-файлы из папки
и объединение из в один тиббл (датафрейм). Предназначено для csv с сайта \href{https://rp5.ru/}{rp5.ru}.
Во всех csv-файлах должны присутствовать столбцы, указанные в \code{col_select} (по умолчанию это
местное время, температура (T), количество осадков (RRR) и высота снежного покрова (sss)).
}
\examples{
\dontrun{
# в самом распространенном случае вам необходимо выполнить
read_rp5_folder("путь_до_вашей_папки")

# при необходимости можно настроить поведение функции:
read_rp5_folder("path/to/your/folder",
                 skip = 6,
                 delim = ";",
                 col_select = c(Local_time = 1,
                                             dplyr::matches('^T$'),
                                             dplyr::matches('^RRR$'),
                                             dplyr::matches('^sss$')),#'
                 encoding = NULL,
                 suppress_warnings = T)  }

}
