% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_date_rp5.R
\name{separate_date_rp5}
\alias{separate_date_rp5}
\title{Выделить год, месяц и день из столбца Local_time в отдельные столбцы.}
\usage{
separate_date_rp5(df, format_of_local_time = "\%d.\%m.\%Y \%R")
}
\arguments{
\item{df}{a tibble: результат выполнения функций \code{\link{read_rp5_csv}} или
\code{\link{read_rp5_folder}}.
Датафрейм обязательно должен содержать столбец Local_time в исходном формате, предоставляемом rp5.}

\item{format_of_local_time}{a character: формат времени, по умолчанию: "\%d.\%m.\%Y \%R".
Подробности: \code{\link[lubridate]{as_date}}}
}
\value{
a tibble: тиббл со столбцами Year (factor), Month (factor), Day (integer), T, RRR, sss (double)
}
\description{
Выделить год, месяц и день из столбца Local_time в отдельные столбцы. Используйте для обработки
датафреймов, полученных с помощью \code{\link{read_rp5_csv}} или \code{\link{read_rp5_folder}}.
}
\examples{
\dontrun{
# в большинстве случаев достаточно запустить
separate_date_rp5(ваш_датафрейм)

# можно настроить формат исходной даты
separate_date_rp5(df,
                  format_of_local_time = "\%d.\%m.\%Y \%R")}
}
