% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregating.R
\name{aggregate_rp5}
\alias{aggregate_rp5}
\title{Агрегация данных о температуре, количестве осадков и высоте снежного покрова}
\usage{
aggregate_rp5(df, resolution)
}
\arguments{
\item{df}{a tibble: тиббл, полученный после выполнения функций \code{\link{read_rp5_csv}} или \code{\link{read_rp5_folder}} и
\code{\link{separate_date_rp5}}. Должен содержать в себе столбцы:
Year, Month, Day - для агрегации до суточного разрешения (работа с сырыми данными);
Year, Month - для агрегации до месячного разрешения (работа с предобработанными данными суточного разрешения);
Year - для агрегации до годового разрешения;
опционально - любой из T, RRR, sss (или T_avg, RRR_sum, sss_avg,
если предварительно агрегировали данные с помощью \code{\link{aggregate_rp5}}) или в любом сочетании.
Остальные столбцы учитываться не будут, даже если они присутствуют в датафрейме.}

\item{resolution}{character: до какого разрешения агрегировать данные? До суточного: resolution = 'day',
месячного: resolution = 'month', годового: resolution = 'year'.}
}
\value{
a tibble: Тиббл со столбцами Year, Month, Day (в зависимости от выбранного разрешения),
N - количество наблюдений в группе (обязательно), и, в зависимости от исходного набора столбцов
T_avg - средняя температура, RRR_sum - сумма осадков, sss_avg - средняя высота снежного покрова.
}
\description{
Вычисляет среднюю температуру, сумму осадков и среднюю высоту снежного покрова для заданного разрешения: годового,
месячного или суточного. Пропущенные значения исключает из расчетов. Работает с тибблом - результатом
последовательного выполнения функций \code{\link{read_rp5_csv}} или \code{\link{read_rp5_folder}} и
\code{\link{separate_date_rp5}}.
}
\examples{
\dontrun{
# Примеры полного пайплайна:
# агрегация до суточного разрешения:
read_rp5_folder('path') |>
  separate_date_rp5() |> # не забудьте разделить компоненты даты
  na_fill_temp_rp5() |>  # опционально, но рекомендуется обработать пропуски
  aggregate_rp5('day')

# агрегация до месячного или годового разрешения
# (рекомендуется предварительно агрегировать до суточного и обработать пропуски)
read_rp5_folder('path') |>
  separate_date_rp5() |> # не забудьте разделить компоненты даты
  na_fill_temp_rp5() |>  # опционально, но рекомендуется обработать пропуски
  aggregate_rp5('month') # можно заменить на 'year'
}
}
