% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_transitions.R
\name{find_transitions}
\alias{find_transitions}
\title{Find the number of parameter transitions through the threshold level}
\usage{
find_transitions(df, parameter, threshold)
}
\arguments{
\item{df}{a tibble containing the columns: year, month, day, parameter (for example, temperature)}

\item{parameter}{a character: the name of the column to search for the number of crossings over the threshold}

\item{threshold}{a numeric: threshold value}
}
\value{
a tibble: Year, Month, Num_of_transitions - the number of parameter transitions through the threshold level
}
\description{
Find the number of parameter transitions through the threshold level
}
\examples{
\dontrun{
counting the number of temperature transitions after 0 degrees
find_transitions(dataframe, T, 0)}
}
