% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rp5.R
\name{read_rp5_folder}
\alias{read_rp5_folder}
\title{Импорт всех csv файлов из папки и объединение их в один тиббл}
\usage{
read_rp5_folder(
  path,
  skip = 6,
  delim = ";",
  local_time_position = 1,
  col_select = c("T", "RRR", "sss"),
  encoding = NULL,
  suppress_warnings = T
)
}
\arguments{
\item{path}{character: путь к папке с csv файлами.}

\item{skip}{integer: то же, что и в \code{\link{read_rp5_csv}}}

\item{delim}{character: то же, что и в \code{\link{read_rp5_csv}}}

\item{local_time_position}{integer: то же, что и в \code{\link{read_rp5_csv}}}

\item{col_select}{character: то же, что и в \code{\link{read_rp5_csv}}}

\item{encoding}{character: то же, что и в \code{\link{read_rp5_csv}}}

\item{suppress_warnings}{bool: то же, что и в \code{\link{read_rp5_csv}}}
}
\value{
tibble со столбцами, указанными в col_select
}
\description{
Обертка над функцией \code{\link{read_rp5_csv}}, позволяющая импортировать сразу все csv-файлы из папки
и объединение из в один тиббл (датафрейм). Предназначено для csv с сайта \href{https://rp5.ru/}{rp5.ru}.
Во всех csv-файлах должны присутствовать столбцы, указанные в \code{col_select} (по умолчанию это
местное время, температура (T), количество осадков (RRR) и высота снежного покрова (sss)).
}
\examples{
\dontrun{
# в самом распространенном случае вам необходимо выполнить
read_rp5_folder("путь_до_вашей_папки")

# при необходимости можно настроить поведение функции:
read_rp5_folder("path/to/your/folder",
                 skip = 6,
                 delim = ";",
                 local_time_position = 1,
                 col_select = c('T', 'RRR', 'sss'),
                 encoding = NULL,
                 suppress_warnings = T)  }

}
